Refactored SQLite implementation into driver
